aaaaaaaaaa stules not found kill me

+++ src/main/java/application/repos/ArticleRepo.java
@@ -3,5 +3,5 @@ package application.repos;
 import application.entities.Article;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ArticleRepo extends CrudRepository<Article, Long> {
+public interface ArticleRepo extends CrudRepository<Article, Integer> {
 }
