+++ src/main/java/application/repos/AuthorRepo.java
@@ -3,5 +3,5 @@ package application.repos;
 import application.entities.Author;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AuthorRepo extends CrudRepository<Author, Long> {
+public interface AuthorRepo extends CrudRepository<Author, Integer> {
 }
